Revert "made a normal pid"
This reverts commit 13be2e1bfa3611fa3051cd2e6ef2da87151cf738.
