+++ src/main/java/frc/robot/commands/VisionPID.java
@@ -141,13 +141,4 @@ public class VisionPID extends Command {
     protected void interrupted() {
         end();
     }
-
-    public void setPID(PidSettings pidSettings){
-        if(!isRunning())
-            pidControllerX.setPID(pidSettings.getKP(),pidSettings.getKI(),pidSettings.getKD());
-    }
-    public void setPID2(PidSettings pidSettings2){
-        if(!isRunning()&&this.isDrivingForward)
-            pidControllerY.setPID(pidSettings2.getKP(),pidSettings2.getKI(),pidSettings2.getKD());
-    }
 }
