+++ src/main/java/frc/robot/motionprofiling/FollowPath.java
@@ -1,7 +1,6 @@
 package frc.robot.motionprofiling;
 
 import edu.wpi.first.wpilibj.command.Command;
-import frc.robot.PidSettings;
 import frc.robot.Robot;
 import frc.robot.RobotConstants;
 import frc.robot.motionprofiling.Path;
@@ -85,11 +84,4 @@ public class FollowPath extends Command {
   protected void interrupted() {
     end();
   }
-  public void setPID(PidSettings pidSettings){
-      this.left.configurePIDVA(pidSettings.getKP(),pidSettings.getKI(),pidSettings.getKD(),pidSettings.getKV(),pidSettings.getKA());
-  }
-  public void setPID2(PidSettings pidSettings){
-    this.right.configurePIDVA(pidSettings.getKP(),pidSettings.getKI(),pidSettings.getKD(),pidSettings.getKV(),pidSettings.getKA());
-  }
-
 }
